resolve #13: Add gtest from conan
* Add one test for long string
* Remove osx build
