--- Sources.cmake
@@ -1,10 +1,0 @@
-set(ALGORITHMS_ON_STRINGS_LIB_SRC
-  algorithms_on_strings/longest_common_substring/LongestSubstring.cpp
-  algorithms_on_strings/longest_common_substring/LongestSubstring.h
-)
-
-set(SORTING_ALGORITHMS_LIB_SRC
-  sorting_algorithms/quick_sort/parallel_and_sequential_quick_sort.cpp
-  sorting_algorithms/quick_sort/qsort_test.cpp
-  sorting_algorithms/quick_sort/template_qsort.h
-)
