+++ conanfile.txt
@@ -1,0 +1,8 @@
+[requires]
+gtest/1.8.1@bincrafters/stable
+
+[options]
+gtest:shared=False
+
+[generators]
+cmake
